workaround for Java code generation for pipelines
- hide the builders checkbox in pipeline export for Java
- force builders = false in transpilation for Java Pipelines
